Extract zxcvbn util
