+++ portal/jest.config.js
@@ -4,8 +4,5 @@ module.exports = {
   transform: {
     "^.+\\.tsx?$": "ts-jest",
   },
-  moduleNameMapper: {
-    "\\.(css|less)$": "identity-obj-proxy",
-  },
   transformIgnorePatterns: ["<rootDir>/node_modules/"],
 };
