+++ portal/package.json
@@ -46,7 +46,6 @@
     "@types/zxcvbn": "4.4.4",
     "eslint": "8.45.0",
     "graphql": "16.8.1",
-    "identity-obj-proxy": "3.0.0",
     "jest": "29.7.0",
     "jest-environment-jsdom": "29.7.0",
     "parcel": "2.10.2",
